# SuperAnswersAngular/src/data-services/models/add-question-answer-request.ts
/* tslint:disable */
export interface AddQuestionAnswerRequest {
  questionId?: string;
  answer?: string;
  createdAt: string;
  createdBy?: string;
  updatedAt: string;
  updatedBy?: string;
  votes: number;
  userId?: string;
}